+++ src/arcadia_microscopy_tools/typing.py
@@ -2,8 +2,9 @@ import numpy as np
 from numpy.typing import NDArray
 
 BoolArray = NDArray[np.bool_]
-FloatArray = NDArray[np.float64]
-Int64Array = NDArray[np.int64]
+UByteArray = NDArray[np.uint8]
 UInt16Array = NDArray[np.uint16]
-UByteArray = NDArray[np.ubyte]
-ScalarArray = BoolArray | FloatArray | Int64Array | UInt16Array
+Int64Array = NDArray[np.int64]
+Float64Array = NDArray[np.float64]
+
+ScalarArray = BoolArray | Float64Array | Int64Array | UInt16Array | UByteArray